backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
langchain>=0.2.11
langchain-core>=0.2.10
langchain-google-genai>=2.0.0
langchain-community>=0.2.5
langchain-huggingface>=0.0.3
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
unstructured>=0.12.6
tiktoken>=0.7.0
huggingface-hub>=0.23.0
pypdf>=4.2.0
beautifulsoup4>=4.12.0
chromadb>=0.4.0
rank_bm25>=0.2.2
flashrank>=0.2.0
langchain-chroma>=0.1.0
